Fix: Refactored code
